% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_distribution.R
\name{calc_chi_sq_stat}
\alias{calc_chi_sq_stat}
\title{Calculate Chi-Squared Test Statistic}
\usage{
calc_chi_sq_stat(obs_freq, exp_freq)
}
\arguments{
\item{obs_freq}{Numeric vector of observed frequencies (O).}

\item{exp_freq}{Numeric vector of expected frequencies (E).}
}
\value{
The Chi-Squared test statistic (numeric).
}
\description{
Calculates the Chi-Squared test statistic: sum( (O - E)^2 / E ).
}
