% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square.R
\name{chi_square_greater}
\alias{chi_square_greater}
\title{Chi-square dist greater}
\usage{
chi_square_greater(n, sd, threshold)
}
\arguments{
\item{n}{number of sample.}

\item{sd}{standard deviation of samples.}

\item{threshold}{number of units.}
}
\value{
A \code{probability} that sample variance is greater than threshold
}
\description{
compute the probability of sample variance is greater than threshold.
}
\examples{
chi_square_greater(25,1200^2,2500)

}
