% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square.R
\name{chi_square_lower}
\alias{chi_square_lower}
\title{Chi-square dist lower}
\usage{
chi_square_lower(n, sd, threshold)
}
\arguments{
\item{n}{number of sample.}

\item{sd}{standard deviation of samples.}

\item{threshold}{number of units.}
}
\description{
compute the probability of sample variance is lower than threshold.
}
