% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{combinatory_number}
\alias{combinatory_number}
\title{function for combinatory numbers}
\usage{
combinatory_number(n, x)
}
\arguments{
\item{n}{total number of items}

\item{x}{number of desired items}
}
\value{
The calculated combinatory number (n choose x)
}
\description{
function for combinatory numbers
}
