% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_score.R
\name{confidence_binomial}
\alias{confidence_binomial}
\title{Confidence_score binomial}
\usage{
confidence_binomial(n, p, z)
}
\arguments{
\item{n}{number of sample.}

\item{p}{probability of success}

\item{z}{value of alpha}
}
\description{
compute the confidence scores.
}
\examples{
confidence_binomial(500,0.42,0.0025)

}
