% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_score.R
\name{confidence_score_t}
\alias{confidence_score_t}
\title{Confidence_score T}
\usage{
confidence_score_t(z, n)
}
\arguments{
\item{z}{value of alpha}

\item{n}{number of sample.}
}
\value{
A \code{score}.
}
\description{
Confidence_score T
}
\examples{
confidence_score_z(0.0025)
}
