% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_score.R
\name{confidence_score_z}
\alias{confidence_score_z}
\title{Confidence_score Z}
\usage{
confidence_score_z(z)
}
\arguments{
\item{z}{value of alpha}
}
\value{
A \code{score}.
}
\description{
Confidence_score Z
}
\examples{
confidence_score_z(0.0025)
}
