% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_distribution.R
\name{estimate_grouped_mean}
\alias{estimate_grouped_mean}
\title{Estimate Mean from Grouped Data}
\usage{
estimate_grouped_mean(x_midpoints, obs_freq)
}
\arguments{
\item{x_midpoints}{Numeric vector of class midpoints.}

\item{obs_freq}{Numeric vector of observed frequencies.}
}
\value{
The estimated mean (numeric).
}
\description{
Calculates the estimated mean (mu_hat) from grouped data using midpoints and frequencies.
}
