% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_distribution.R
\name{estimate_grouped_sd}
\alias{estimate_grouped_sd}
\title{Estimate Standard Deviation from Grouped Data}
\usage{
estimate_grouped_sd(x_midpoints, obs_freq)
}
\arguments{
\item{x_midpoints}{Numeric vector of class midpoints.}

\item{obs_freq}{Numeric vector of observed frequencies.}
}
\value{
The estimated sample standard deviation (numeric).
}
\description{
Calculates the estimated sample standard deviation (sig_hat) from grouped data.
}
