% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_student.R
\name{t_student}
\alias{t_student}
\title{T student distribution}
\usage{
t_student(n, sd, difference)
}
\arguments{
\item{n}{number of sample.}

\item{sd}{standard deviation of samples.}

\item{difference}{number of units.}
}
\value{
A \code{probability} that the sample mean doesn't differ more than difference unit.
}
\description{
compute the probability of sample not being more deviated than difference.
}
\examples{
t_student(16,10.85,8)

}
