% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_distribution.R
\name{p_normal_interval}
\alias{p_normal_interval}
\alias{p_normal_lower_equal
Probability of X being greater than q}
\alias{p_normal_greater}
\alias{p_normal_lower_equal}
\alias{p_normal_lower}
\title{Normal Distribution N(mean, sd)}
\usage{
p_normal_interval(a, b, mean, sd)

p_normal_greater(x, mean, sd)

p_normal_lower_equal(x, mean, sd)

p_normal_lower(x, mean, sd)
}
\arguments{
\item{a}{Lower bound of the interval.}

\item{b}{Upper bound of the interval.}

\item{mean}{Numeric population mean (mu).}

\item{sd}{Positive numeric population standard deviation (sigma).}

\item{x}{Numeric value(s) at which to evaluate the probability function (e.g., the threshold q).}
}
\value{
Numeric vector of probabilities.
}
\description{
cumulative probability functions for the Normal distribution with parameters mean (mu) and sd (sigma).
}
\details{
The Normal distribution (Gaussian distribution) is a continuous distribution modeling data that clusters around a central mean.
}
\concept{continuous-distributions}
